refactor(pdf): isolate source math provenance
